# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TrustCart AI+ API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication
  - name: products
    description: Product management
  - name: cart
    description: Shopping cart
  - name: wishlist
    description: Wishlist
  - name: orders
    description: Orders
  - name: preferences
    description: User preferences
  - name: recommendations
    description: AI recommendations
  - name: vault
    description: SecureSync Vault
  - name: dashboard
    description: AI Dashboard analytics

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/register:
    post:
      operationId: registerUser
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/login:
    post:
      operationId: loginUser
      tags: [auth]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/logout:
    post:
      operationId: logoutUser
      tags: [auth]
      summary: Logout user
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated

  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products with filters
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: mood
          in: query
          schema:
            type: string
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: minRating
          in: query
          schema:
            type: number
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Product list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"

  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found

  /products/{id}/related:
    get:
      operationId: getRelatedProducts
      tags: [products]
      summary: Get related products
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Related products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"

  /recommendations:
    get:
      operationId: getRecommendations
      tags: [recommendations]
      summary: Get AI personalized recommendations
      parameters:
        - name: mood
          in: query
          schema:
            type: string
        - name: userId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Recommended products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"

  /cart:
    get:
      operationId: getCart
      tags: [cart]
      summary: Get cart items
      responses:
        "200":
          description: Cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    post:
      operationId: addToCart
      tags: [cart]
      summary: Add item to cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddToCartBody"
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"

  /cart/{itemId}:
    delete:
      operationId: removeFromCart
      tags: [cart]
      summary: Remove item from cart
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"

  /cart/clear:
    delete:
      operationId: clearCart
      tags: [cart]
      summary: Clear cart
      responses:
        "200":
          description: Cleared
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  /wishlist:
    get:
      operationId: getWishlist
      tags: [wishlist]
      summary: Get wishlist
      responses:
        "200":
          description: Wishlist
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: addToWishlist
      tags: [wishlist]
      summary: Add item to wishlist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WishlistBody"
      responses:
        "200":
          description: Updated wishlist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  /wishlist/{productId}:
    delete:
      operationId: removeFromWishlist
      tags: [wishlist]
      summary: Remove from wishlist
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List user orders
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create order from cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderBody"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  /preferences:
    get:
      operationId: getPreferences
      tags: [preferences]
      summary: Get user preferences
      responses:
        "200":
          description: Preferences
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserPreferences"
    put:
      operationId: updatePreferences
      tags: [preferences]
      summary: Update preferences
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePreferencesBody"
      responses:
        "200":
          description: Updated preferences
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserPreferences"

  /vault:
    get:
      operationId: getVaultStatus
      tags: [vault]
      summary: Get vault / privacy status
      responses:
        "200":
          description: Vault status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VaultStatus"

  /vault/toggle:
    post:
      operationId: toggleVaultMode
      tags: [vault]
      summary: Toggle anonymous/personalized mode
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VaultToggleBody"
      responses:
        "200":
          description: Updated vault status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VaultStatus"

  /vault/clear:
    delete:
      operationId: clearVaultData
      tags: [vault]
      summary: Clear all user data from vault
      responses:
        "200":
          description: Cleared
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  /dashboard/stats:
    get:
      operationId: getDashboardStats
      tags: [dashboard]
      summary: Get dashboard analytics summary
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"

  /dashboard/accuracy:
    get:
      operationId: getRecommendationAccuracy
      tags: [dashboard]
      summary: Get recommendation accuracy over time
      responses:
        "200":
          description: Accuracy data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AccuracyDataPoint"

  /dashboard/category-breakdown:
    get:
      operationId: getCategoryBreakdown
      tags: [dashboard]
      summary: Get category preference breakdown
      responses:
        "200":
          description: Category data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CategoryBreakdown"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]

    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]

    RegisterBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        name:
          type: string
      required: [email, password, name]

    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]

    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required: [user, token]

    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        createdAt:
          type: string
      required: [id, email, name, createdAt]

    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        price:
          type: number
        imageUrl:
          type: string
        category:
          type: string
        mood:
          type: string
        rating:
          type: number
        reviewCount:
          type: integer
        trustScore:
          type: number
        regretPercent:
          type: number
        colors:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
        aiReason:
          type: string
        isLocalData:
          type: boolean
        inStock:
          type: boolean
      required: [id, name, price, imageUrl, category, rating, trustScore, regretPercent, inStock]

    Cart:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
        total:
          type: number
        itemCount:
          type: integer
      required: [items, total, itemCount]

    CartItem:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        product:
          $ref: "#/components/schemas/Product"
        quantity:
          type: integer
        price:
          type: number
      required: [id, productId, product, quantity, price]

    AddToCartBody:
      type: object
      properties:
        productId:
          type: integer
        quantity:
          type: integer
      required: [productId]

    WishlistBody:
      type: object
      properties:
        productId:
          type: integer
      required: [productId]

    Order:
      type: object
      properties:
        id:
          type: integer
        status:
          type: string
        total:
          type: number
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
        address:
          type: string
        createdAt:
          type: string
      required: [id, status, total, items, address, createdAt]

    CreateOrderBody:
      type: object
      properties:
        address:
          type: string
        city:
          type: string
        state:
          type: string
        zipCode:
          type: string
        country:
          type: string
      required: [address, city, state, zipCode, country]

    UserPreferences:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        favoriteColors:
          type: array
          items:
            type: string
        favoriteCategories:
          type: array
          items:
            type: string
        currentMood:
          type: string
        privacyMode:
          type: string
      required: [id, userId, favoriteColors, favoriteCategories, currentMood, privacyMode]

    UpdatePreferencesBody:
      type: object
      properties:
        favoriteColors:
          type: array
          items:
            type: string
        favoriteCategories:
          type: array
          items:
            type: string
        currentMood:
          type: string
        privacyMode:
          type: string

    VaultStatus:
      type: object
      properties:
        isAnonymous:
          type: boolean
        hashedUserId:
          type: string
        encryptionStatus:
          type: string
        localStorageSize:
          type: integer
        cloudSyncEnabled:
          type: boolean
        lastSyncAt:
          type: string
        dataCategories:
          type: array
          items:
            type: string
      required: [isAnonymous, hashedUserId, encryptionStatus, localStorageSize, cloudSyncEnabled, dataCategories]

    VaultToggleBody:
      type: object
      properties:
        anonymous:
          type: boolean
      required: [anonymous]

    DashboardStats:
      type: object
      properties:
        totalRecommendations:
          type: integer
        acceptedRecommendations:
          type: integer
        accuracyRate:
          type: number
        totalProducts:
          type: integer
        activeUsers:
          type: integer
        avgTrustScore:
          type: number
        topMood:
          type: string
        privacyCompliance:
          type: number
      required: [totalRecommendations, acceptedRecommendations, accuracyRate, totalProducts, activeUsers, avgTrustScore, topMood, privacyCompliance]

    AccuracyDataPoint:
      type: object
      properties:
        date:
          type: string
        accuracy:
          type: number
        recommendations:
          type: integer
      required: [date, accuracy, recommendations]

    CategoryBreakdown:
      type: object
      properties:
        category:
          type: string
        count:
          type: integer
        percentage:
          type: number
      required: [category, count, percentage]