# lib/api-zod/src/generated/types/createOrderBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TrustCart AI+ API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateOrderBody {
  address: string;
  city: string;
  state: string;
  zipCode: string;
  country: string;
}