# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TrustCart AI+ API
 * OpenAPI spec version: 0.1.0
 */
import type { CartItem } from "./cartItem";

export interface Order {
  id: number;
  status: string;
  total: number;
  items: CartItem[];
  address: string;
  createdAt: string;
}