# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TrustCart AI+ API
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  name: string;
  description?: string;
  price: number;
  imageUrl: string;
  category: string;
  mood?: string;
  rating: number;
  reviewCount?: number;
  trustScore: number;
  regretPercent: number;
  colors?: string[];
  tags?: string[];
  aiReason?: string;
  isLocalData?: boolean;
  inStock: boolean;
}